moved some sqrt utils
