+++ src/probnum/filtsmooth/gaussfiltsmooth/__init__.py
@@ -3,8 +3,8 @@ from .iterated_kalman import IteratedKalman
 from .kalman import Kalman
 from .kalmanposterior import KalmanPosterior
 from .linearizing_transition import LinearizingTransition
-from .square_root_kalman import SquareRootKalman
+from .sqrt_kalman import SquareRootKalman
+from .sqrt_utils import cholesky_update
 from .stoppingcriterion import FixedPointStopping, StoppingCriterion
 from .unscentedkalman import ContinuousUKFComponent, DiscreteUKFComponent, UKFComponent
 from .unscentedtransform import UnscentedTransform
-from .utils import cholesky_update
